refactor : remove HomeScreen and integrate to ProjectApp - homescreen의 의미가 애매해서 projectApp으로 옮김 - DashboardScreen에 있는 Scaffold 하나로 합침 - projectAppState에 scaffoldState 추가
